requirements.txt
aiofiles==0.4.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
decorator==4.4.0
defusedxml==0.6.0
Django==2.2.1
django-extensions==2.1.6
django-extra-fields==1.2.1
django-filters==0.2.1
djangorestframework==3.9.4
entrypoints==0.3
Flask==1.0.2
httptools==0.0.13
idna==2.8
ipykernel==5.1.0
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
MarkupSafe==1.1.1
mistune==0.8.4
multidict==4.5.2
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
numpy==1.16.3
pandas==0.24.2
pandocfilters==1.4.2
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
pkg-resources==0.0.0
prometheus-client==0.6.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.0
pyrsistent==0.15.2
python-dateutil==2.8.0
pytz==2019.1
pyzmq==18.0.1
qtconsole==4.4.4
requests==2.21.0
sanic==19.3.1
Send2Trash==1.5.0
six==1.12.0
sqlparse==0.3.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.2
traitlets==4.3.2
ujson==1.35
urllib3==1.24.3
uvloop==0.12.2
wcwidth==0.1.7
webencodings==0.5.1
websockets==6.0
Werkzeug==0.15.2
widgetsnbextension==3.4.2
